feat: add changing header's color when scrollTop
